Menuing for replays.

+++ Libraries/Brawl/SC/scMelee.cpp
@@ -4,6 +4,6 @@
 
 #include "scMelee.h"
 
-u8 stOperatorReadyGo::isEnd() {
+unsigned int stOperatorReadyGo::isEnd() {
     return _isEnd_ST_OPERATOR_READY_GO(this);
 }
